Issue 1756: Move CLI service methods into Service class with Transactional.
The `@Transactional` does not consistently work when run under the Cli Component.
Sometimes the Submissions appear but most often times they do not.

Move the code into a separate class file with `@Transactional` and `@Service` and then calling those methods appears to produce the expected behavior.
